fix(toast): Make placement a config setting
