+++ packages/ng-primitives/toast/src/toast/toast-manager.ts
@@ -12,7 +12,7 @@ import {
 } from '@angular/core';
 import { createPortal, NgpPortal } from 'ng-primitives/portal';
 import { injectToastConfig } from '../config/toast-config';
-import { NgpToast, NgpToastSwipeDirection } from './toast';
+import { NgpToast, NgpToastPlacement, NgpToastSwipeDirection } from './toast';
 import { provideToastContext } from './toast-context';
 import { provideToastOptions } from './toast-options';
 
@@ -39,7 +39,7 @@ export class NgpToastManager {
     const viewContainerRef = this.applicationRef.components[0].injector.get(ViewContainerRef);
 
     let instance: NgpToast | null = null;
-    const placement = options.placement ?? 'top-end';
+    const placement = options.placement ?? this.config.placement;
     const container = this.getOrCreateContainer(placement);
 
     const portal = createPortal(
@@ -178,14 +178,6 @@ export class NgpToastManager {
     return container;
   }
 }
-
-export type NgpToastPlacement =
-  | 'top-start'
-  | 'top-end'
-  | 'top-center'
-  | 'bottom-start'
-  | 'bottom-end'
-  | 'bottom-center';
 
 export interface NgpToastOptions<T = unknown> {
   /** The position of the toast */
